use i64/u64 instead of isize/usize
fixes https://github.com/ayrat555/frankenstein/issues/10

+++ src/api.rs
@@ -571,7 +571,7 @@ pub trait TelegramApi {
     fn get_chat_members_count(
         &self,
         params: &GetChatMembersCountParams,
-    ) -> Result<MethodResponse<usize>, Self::Error> {
+    ) -> Result<MethodResponse<u64>, Self::Error> {
         self.request("getChatMembersCount", Some(params))
     }
